Big Update: Code redesign.
Ok, I'm back and started messing with the code again. I'll be modifying code that keeps everything the same in terms of gameplay, but makes everything more versatile in the long run.

This update restructures code and focuses on the courses menu. Now when adding a new course it automatically arranges the items as needed.