WIP: Code Redesign - Course Details Menu
This update focus on the course details menu.

It now scales better and allows for dynamic height increase depending on the amount of content in the menu.

Additionally it now allows the addition checkpoints by reading out of the course menu. (uff it was quite a hassle, but it's done)

NOTE: the OK button to leave the menus is not currently working. I'll be implementing a event system to deal with that in a more dynamic way.